Add files via upload
This program is personally one of my best works whilst programming. The assigment given was to simply enter 10 numbers and sort them from lowest to highest. What i had done here were firstly load the numbers in RAM and use two pointer variables. One varibale to point at the starting address and the other points to the next address. The address contents would be retrieved and compared and if the current adres value is greater than the next value then the context would be switched until the data stored in RAM is sorted. The idea behind this program is exactly the same as the bubblesort algorithm found in the Structured Programming in C repository.